+++ services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/modules/long_running_tasks_utils.py
@@ -9,7 +9,7 @@ from ..core.errors import (
     ContainerExecTimeoutError,
 )
 from ..models.shared_store import SharedStore
-from ..modules.container_utils import run_command_in_container
+from .container_utils import run_command_in_container
 
 _logger = logging.getLogger(__name__)
